🔧 address PR review feedback: extract shared fixture, fix shortcuts, remove log spam
- Extract setupVaultPage() helper into tests/test-helpers.ts
- Remove console.log from waitForFunction polling predicate
- Use page.keyboard.press('l') instead of direct function call for label toggle
- Blur canvas before keyboard shortcuts to avoid focus issues
- Use OS-agnostic search shortcut (Meta+k on macOS, Control+k elsewhere)
