+++ apps/web/tests/vault/initialization.spec.ts
@@ -1,34 +1,9 @@
 import { test, expect } from "@playwright/test";
+import { setupVaultPage } from "../test-helpers";
 
 test.describe("Vault Initialization", () => {
   test.beforeEach(async ({ page }) => {
-    await page.addInitScript(() => {
-      (window as any).DISABLE_ONBOARDING = true;
-      (window as any).__E2E__ = true;
-      try {
-        localStorage.setItem("codex_skip_landing", "true");
-      } catch {
-        /* ignore */
-      }
-    });
-
-    await page.goto("/");
-    // Wait for vault initialization (OPFS auto-load)
-    await page.waitForFunction(
-      () => {
-        const status = (window as any).vault?.status;
-        console.log(
-          `[E2E Wait] Current vault status: ${status}, isInitialized: ${(window as any).vault?.isInitialized}`,
-        );
-        return status === "idle";
-      },
-      {
-        timeout: 15000,
-      },
-    );
-    await expect(page.getByTestId("graph-canvas")).toBeVisible({
-      timeout: 10000,
-    });
+    await setupVaultPage(page);
   });
 
   test("auto-initializes default vault", async ({ page }) => {
